refactor(db): extract chat contact repository
